New Mobile UI changes.

+++ TextPortCore/Models/TimeZone.cs
@@ -4,7 +4,7 @@ namespace TextPortCore.Models
 {
     public partial class TimeZone
     {
-        public short TimeZoneId { get; set; }
+        public int TimeZoneId { get; set; }
         public decimal Utcoffset { get; set; }
         public string Name { get; set; }
     }
